making specialized multiplications maliciously secure
