+++ src/protocol/modulus_conversion/mod.rs
@@ -1,2 +1,2 @@
 pub mod convert_shares;
-mod specialized_mul;
+//mod specialized_mul;
